refactor with sensible names & finish API methods & complete cross language docs
