+++ src/main/java/org/openjfx/starea/DayData.java
@@ -1,0 +1,7 @@
+package org.openjfx.starea;
+
+public class DayData {
+    double apparentTemperatureMin; // C or F
+    double apparentTemperatureMax; // C or F
+    int weatherCode;
+}
