+++ src/main/java/org/openjfx/starea/HourData.java
@@ -1,0 +1,10 @@
+package org.openjfx.starea;
+
+public class HourData {
+    double temperature2m; // C or F
+    int weatherCode;
+    int visibility; // m
+    int precipitationProbability; // %
+    boolean isDay; // bool
+    int cloudCover; // %
+}
