+++ src/main/resources/org/openjfx/starea/index.html
@@ -3,19 +3,15 @@
     <title>Starea</title>
 
     <script>
-        function toggleShape() {
-            var box = document.getElementById("box")
-            var br = box.style.borderRadius
-            box.textContent = br
-            if (br == "0px") { box.style.borderRadius = "10px"; }
-            else { box.style.borderRadius = "0px"; }
+        function func3() {
+            window.remote.log("run func3 in html from java");
+            func4();
         }
-        console.log("alksndklabdn");
     </script>
 </head>
 <body>
     <div class="navbar"><span class="title">Starea</span></div>
-    <button onClick="changeColour()">Click Me!</button>
+    <button onClick="func1()">Click Me!</button>
     <div id="box"></div>
 </body>
 </html>
