+++ src/main/resources/org/openjfx/starea/script.js
@@ -1,10 +1,8 @@
-function changeColour() {
-    var letters = "0123456789ABCDEF"
-    var colour = "#"
-    for (var i = 0; i < 6; i++) {
-        colour += letters[Math.floor(Math.random() * 16)];
-    }
-    var box = document.getElementById("box")
-    box.style.backgroundColor = colour;
-    console.log(colour);
+function func1() {
+    window.remote.log("run func1 in js from button");
+    window.remote.func2();
+}
+
+function func4() {
+    window.remote.log("run func4 in js from html");
 }
